Fixed 3dcorr error with device set as cpu
RuntimeError: "compute_columns3d" not implemented for 'Half'
https://blog.csdn.net/qq_45708644/article/details/135225291
conv3d on cpu is only implemented for float32/64

+++ method/ee3p3d.py
@@ -108,8 +108,8 @@ class EE3P3D:
             template = partitioned_events[:template_depth, :, :, y_win, x_win]
             window = partitioned_events[:, :, :, y_win, x_win]
 
-            corr_out = F.conv3d(window[None][None],
-                                template[None][None]).squeeze()
+            corr_out = F.conv3d(window[None][None].float(),
+                                template[None][None].float()).squeeze()
 
             if torch.topk(corr_out, 2)[0][-1] <= 0:
                 continue
